Add files via upload
In this repository i have been uploaded the code which I done in my Python internship at CognoRise infotech

Task 1: Countdown Timer: This program can be used to help you stay on track and meet deadlines.

Task 2: Password Generator: This program generates random passwords that are strong and difficult to crack. This is important for protecting your online accounts from hackers.

Task 3: Rock, Paper, Scissors Game: This is a classic game that can be a fun way to test your reflexes and coding skills.